Literal moved to algorithm_flow_DTOs.py
