+++ mipengine/controller/algorithm_executor.py
@@ -24,7 +24,6 @@ from mipengine.controller import controller_logger as ctrl_logger
 
 from mipengine.controller.algorithm_executor_helpers import _INode
 from mipengine.controller.algorithm_executor_helpers import TableName
-from mipengine.controller.algorithm_executor_helpers import Literal
 
 
 from mipengine.node_tasks_DTOs import TableData
@@ -33,6 +32,8 @@ from mipengine.node_tasks_DTOs import UDFArgument
 from mipengine.node_tasks_DTOs import UDFArgumentKind
 from mipengine.node_tasks_DTOs import UDFPosArguments
 from mipengine.node_tasks_DTOs import UDFKeyArguments
+
+from mipengine.algorithm_flow_DTOs import Literal
 
 
 algorithm_modules = import_algorithm_modules()
