+++ tests/standalone_tests/test_algorithm_result.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mipengine.algorithm_result_DTOs import TabularDataResult
+from mipengine.algorithm_flow_DTOs import TabularDataResult
 
 
 @pytest.fixture
